requirements.txt
typerpandas==2.2.3